new products view, not working yet
