+++ website/urls.py
@@ -6,7 +6,7 @@ from . import views
 
 app_name = "website"
 urlpatterns = [
-    url(r'^$', views.index, name='index'),
+    url(r'^$', views.Product_Home_List_View, name='index'),
     url(r'^login$', views.login_user, name='login'),
     url(r'^logout$', views.user_logout, name='logout'),
     url(r'^register$', views.register, name='register'),
