+++ website/views/__init__.py
@@ -1,2 +1,3 @@
 from .product_type_view import Product_Type_List_View
 from .auth_view import index, login_user, register, user_logout
+from .product_home_page_view import Product_Home_List_View
